backend/requirements.txt
fastapi
uvicorn
requests
httpx
beautifulsoup4
textstat
reportlab
sqlalchemy
python-multipart
python-dotenv